img_description_server/requirements.txt
flask
pillow
openai
pillow
# pip install -r requirements.txt